moved pairing of vmaf scores and resolutions/bitrates to own function
