+++ src/suggest-ladder.ts
@@ -2,6 +2,7 @@ import { Resolution } from './models/resolution';
 import logger from './logger';
 import { BitrateResolutionVMAF } from './models/bitrate-resolution-vmaf';
 import { getVmaf } from '.';
+import { pairVmafWithResolutionAndBitrate } from './pairVmaf';
 
 /**
  * Suggests an optimal ABR-ladder from a directory of VMAF-files. Only supports loading from S3 at the moment.
@@ -17,31 +18,32 @@ export default async function suggestLadder(
   includeAllBitrates: boolean = false,
   onProgress: (index: number, filename: string, total: number) => void = () => {}
 ): Promise<BitrateResolutionVMAF[]> {
-  let pairs = new Map<number, { resolution: Resolution; vmaf: number }[]>();
+  const pairs = await pairVmafWithResolutionAndBitrate(directoryWithVmafFiles, filterFunction, onProgress);
+  // let pairs = new Map<number, { resolution: Resolution; vmaf: number }[]>();
   let optimal: { resolution: Resolution; vmaf: number; bitrate: number }[] = [];
 
-  logger.info('Loading VMAF data...');
-  const vmafs = await getVmaf(directoryWithVmafFiles, onProgress);
-  let counter = 1;
-  vmafs.forEach(({ filename, vmaf }) => {
-    const [resolutionStr, bitrateStr] = filename.split('_');
-    const [widthStr, heightStr] = resolutionStr.split('x');
+  // logger.info('Loading VMAF data...');
+  // const vmafs = await getVmaf(directoryWithVmafFiles, onProgress);
+  // let counter = 1;
+  // vmafs.forEach(({ filename, vmaf }) => {
+  //   const [resolutionStr, bitrateStr] = filename.split('_');
+  //   const [widthStr, heightStr] = resolutionStr.split('x');
 
-    const width = parseInt(widthStr);
-    const height = parseInt(heightStr);
-    const bitrate = parseInt(bitrateStr);
+  //   const width = parseInt(widthStr);
+  //   const height = parseInt(heightStr);
+  //   const bitrate = parseInt(bitrateStr);
 
-    if (filterFunction(bitrate, { width, height }, vmaf)) {
-      if (pairs.has(bitrate)) {
-        pairs.get(bitrate)?.push({ resolution: { width, height }, vmaf });
-      } else {
-        pairs.set(bitrate, [{ resolution: { width, height }, vmaf }]);
-      }
-    }
+  //   if (filterFunction(bitrate, { width, height }, vmaf)) {
+  //     if (pairs.has(bitrate)) {
+  //       pairs.get(bitrate)?.push({ resolution: { width, height }, vmaf });
+  //     } else {
+  //       pairs.set(bitrate, [{ resolution: { width, height }, vmaf }]);
+  //     }
+  //   }
 
-    logger.info(`Finished loading VMAF ${counter}/${vmafs.length}.`);
-    counter += 1;
-  });
+  //   logger.info(`Finished loading VMAF ${counter}/${vmafs.length}.`);
+  //   counter += 1;
+  // });
 
   // Get optimal resolution for each bitrate
   pairs.forEach((bitrateVmafPairs, bitrate) => {
